Refactored Intcode to support larger values and relative parameter mode

+++ src/main/java/nl/erends/advent/year2019/Day02.java
@@ -3,14 +3,14 @@ package nl.erends.advent.year2019;
 import nl.erends.advent.util.AbstractProblem;
 import nl.erends.advent.util.Util;
 
-public class Day02 extends AbstractProblem<String, Integer> {
+public class Day02 extends AbstractProblem<String, Long> {
 
     public static void main(String[] args) {
         new Day02().setAndSolve(Util.readLine(2019, 2));
     }
     
     @Override
-    public Integer solve1() {
+    public Long solve1() {
         Intcode intcode = new Intcode(input);
         intcode.setCode(1, 12);
         intcode.setCode(2, 2);
@@ -19,7 +19,7 @@ public class Day02 extends AbstractProblem<String, Integer> {
     }
     
     @Override
-    public Integer solve2() {
+    public Long solve2() {
         for (int noun = 0; noun <= 99; noun++) {
             for (int verb = 0; verb <= 99; verb++) {   
                 Intcode intcode = new Intcode(input);
@@ -27,7 +27,7 @@ public class Day02 extends AbstractProblem<String, Integer> {
                 intcode.setCode(2, verb);
                 intcode.execute();
                 if (intcode.getCode(0) == 19690720) {
-                    return 100 * noun + verb;
+                    return 100L * noun + verb;
                 }
             }
         }
